refactor: move IA metrics collector
