+++ test/noyau/unit/ia_metrics_collector.g_test.dart
@@ -1,7 +1,7 @@
 // Copilot Prompt : Test automatique généré pour ia_metrics_collector.g.dart (unit)
 import 'package:flutter_test/flutter_test.dart';
 import '../../test_config.dart';
-import 'package:anisphere/modules/noyau/logic/ia_metrics_collector.dart';
+import 'package:anisphere/services/ia/ia_metrics_collector.dart';
 
 void main() {
   setUpAll(() async {
